feat(user-deletion): add functionality to delete user forms and responses
This commit introduces new methods to handle the deletion of user-related data, including forms and responses, when a user is deleted. This ensures that all associated data is properly cleaned up, maintaining data integrity and consistency in the system.

+++ module/services/user.service.ts
@@ -1,12 +1,22 @@
 import { IUserFromDB } from "@/@types";
 import userRepo from "../repositories/user.repo";
+import formsRepo from "../repositories/forms.repo";
+import responseRepo from "../repositories/response.repo";
 
 class UserService {
     async findUserById(id: string) {
         return await userRepo.getUserById(id);
     }
     async deleteUser(userId: string) {
+        const user = await userRepo.getUserById(userId);
+        if(!user) {
+            throw new Error("User not found");
+        }
+
         const deletedUser: IUserFromDB | null = await userRepo.deleteUser(userId);
+        if(!deletedUser) {
+            throw new Error("Error deleting user");
+        }
         return deletedUser;
     }
 }
